requirements.txt
Django>=6.0
asgiref
sqlparse
tzdata
playwright
pytest-playwright
pytest-django
